% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_ci.R
\name{kappa_ci}
\alias{kappa_ci}
\title{Cohen's Kappa Confidence Interval}
\usage{
kappa_ci(x, level = 0.95)
}
\arguments{
\item{x}{A table.}

\item{level}{Level of the confidence interval.}
}
\value{

}
\description{
Cohen's Kappa Confidence Interval
}
