% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_ci.R
\name{kappa_ci}
\alias{kappa_ci}
\title{Cohen's Kappa Confidence Interval}
\usage{
kappa_ci(x, level = 0.95)
}
\arguments{
\item{x}{A 2x2 table.}

\item{level}{Level of the confidence interval.}
}
\description{
Implements the exact bootstrap confidence interval for Cohen's kappa in small
samples by Klar et al. (2002).
}
\examples{
tab <- matrix(c(10, 2, 2, 4), ncol = 2)
kappa_ci(tab)
}
